requirements.txt
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-anthropic
sentence-transformers
chromadb
anthropic
streamlit
pypdf
python-dotenv
